requirements.txt
Django==1.10
django-cleanup==0.4.2
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
python-magic==0.4.12
sorl-thumbnail==12.3
